style: changed request organisation, added tests for get operation
