+++ serenity-rest-assured/src/test/groovy/net/serenitybdd/rest/staging/WhenExecutingGetRequest.groovy
@@ -2,14 +2,12 @@ package net.serenitybdd.rest.staging
 
 import com.github.tomakehurst.wiremock.client.WireMock
 import com.github.tomakehurst.wiremock.junit.WireMockRule
-import com.jayway.restassured.RestAssured
 import com.jayway.restassured.specification.RequestSender
-import net.serenitybdd.rest.staging.decorators.RequestSpecificationDecorated
+import net.serenitybdd.rest.staging.decorators.request.RequestSpecificationDecorated
 import net.serenitybdd.rest.staging.decorators.ResponseDecorated
 import net.serenitybdd.rest.staging.decorators.ResponseSpecificationDecorated
 import net.serenitybdd.rest.staging.rules.RestConfigurationAction
 import net.serenitybdd.rest.staging.rules.RestConfigurationRule
-import org.hamcrest.Matchers
 import org.junit.Rule
 import spock.lang.Specification
